requirements.txt
streamlit
pyodbc
pandas
numpy
matplotlib
python-dotenv
